add if(Not)Exists option to S3 bucket tasks.

+++ src/main/groovy/jp/classmethod/aws/gradle/s3/BucketTask.groovy
@@ -1,5 +1,6 @@
 package jp.classmethod.aws.gradle.s3
 
+import com.amazonaws.*;
 import com.amazonaws.services.s3.*
 import com.amazonaws.services.s3.model.*
 import com.amazonaws.services.s3.transfer.*
@@ -18,13 +19,27 @@ public class CreateBucketTask extends DefaultTask {
 	
 	def String bucketName
 	
+	def boolean ifNotExists
+	
 	@TaskAction
 	def createBucket() {
 		if (! bucketName) throw new GradleException("bucketName is not specified")
 		
 		def AmazonS3Client s3 = project.aws.s3
-		s3.createBucket(bucketName)
-		println "bucket $bucketName created"
+		if (ifNotExists == false || exists() == false) {
+			s3.createBucket(bucketName)
+			println "bucket $bucketName created"
+		}
+	}
+	
+	def boolean exists() {
+		def AmazonS3Client s3 = project.aws.s3
+		try {
+			s3.getBucketLocation(bucketName)
+			true
+		} catch (AmazonClientException e) {
+			false
+		}
 	}
 }
 
@@ -37,12 +52,26 @@ public class DeleteBucketTask extends DefaultTask {
 	
 	def String bucketName
 	
+	def boolean ifExists
+	
 	@TaskAction
 	def deleteBucket() {
 		if (! bucketName) throw new GradleException("bucketName is not specified")
 		
 		def AmazonS3Client s3 = project.aws.s3
-		s3.deleteBucket(bucketName)
-		println "bucket $bucketName deleted"
+		if (ifExists == false || exists()) {
+			s3.deleteBucket(bucketName)
+			println "bucket $bucketName deleted"
+		}
+	}
+	
+	def boolean exists() {
+		def AmazonS3Client s3 = project.aws.s3
+		try {
+			s3.getBucketLocation(bucketName)
+			true
+		} catch (AmazonClientException e) {
+			false
+		}
 	}
 }
